requirements.txt
Django==2.2.13 
djangorestframework==3.9.1
bcrypt==3.1.7 
gunicorn == 20.0.4
django-heroku==0.3.1
django-cors-headers==3.2.1
PyJWT==1.7.1 
coverage==5.1
# psycopg2==2.8.4 
# Pillow==7.2.0
python-decouple==3.3 
requests==2.24.0
cloudinary==1.22.0